requirements.txt
telethon>=1.36.0
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
